Updates to the code that make it more functional
I am still not done with debugging the code, but these are some updates to get it working further. Honestly the change in FeatureExporter.py might not be necessary, but I think helps anyways. The other changes add event_type directly from the row's event_custom.

The next bug to tackle is just the refactor feature names to match the new names.