Lakeland Update
Not yet debugged, but now it calculates ~2/3 of the features, including aggregates and per-windows. Runtime speed still also hasn't been tested and refactoring is necessary because the code is messy. Major updates include logging emotes, switching item mark logs to be based on the tiles and not the corn, aggregates from the perwindow options, and aggregates based on "secs to" certain checkpoints