Update to Lakeland Feature Exporter
This update contains many more features calculated in LakelandExtractor.py and an update to the LAKELAND.json schema, but also contains two changes which may affect other extractors. I will have to talk to Luke about them.

Mainly, GameTables now include a GameTable.playtimes which includes the length of all the playtimes. If the game_id is LAKELAND, then the min level and max level is now the max_play_duration / WINDOW_SIZE_SECONDS which can be found in config.json.template.

It's still a little messy, but I don't think it should break anything.