Better handling of ssh/sql objects, and initial work to export all data for a game.
We no longer have ExportManager keep an sql connection,
just open and close connections as needed.
Maybe not as nice as having a steady line to sql, but
this way other code can't accidentally close the manager's connection.

Also, started to add some code to export all months of
a game's data. Still a work in progress, as we ran into
database issues.