Reworked logging code.
Previously, I'd started passing a couple loggers everywhere.
That led to lots of errors with missed params, and wasn't a
clean design either.
Now, I've just got a global module in utils used to log,
either to stdout or file. Should be a nicer thing,
and all the main work is done in one commit,
instead of being spread over multiple commits.

So, any direct follow-ups should be bug-fixes, no 'in-progress' stuff.