Rename previously exported files before re-exporting.
Essentially, whatever files are indexed for the site,
we rename to whatever new file name is used on a
new export with the same data id.
This way, we don't end up with a bunch of extra
outdated files sitting around on the server.

Also, re-organized a couple things so the code is
easier to read after adding this feature.