JowilderExtractor: Notebook/Wildcard Fix 2
Duplicates were Q10 and Q16. It turns out that in both of those cases, the user seems to be able to select multiple answers (the answers are right next to each other) and my guess is that only the first answer gets checked whereas the second answer is actually the correct one. Because the logger doesn’t go directly through the wildcard code itself (because that code is generated from separate files), it does not calculate right/wrong in the same way. So now I just go through and change any answer which appeared to be a mistake into ?’s. It can probably be assumed that ?’s in Q10 are either E or F and ?’s in Q16 are  either c or d.