Lakeland Extractor: Add "event_sequence", remove dead feats, fix bugs
The bug fix was in "item marked..." and "per capita item marked..." type features. Previously, only the "marked_use" column had values, as the result of was a result of self._tiles not updating with the appropriate usage marks. Now it is fixed so that poop_marked_sell , corn_marked_eat_per_capita, etc shows numbers instead of all 0/null.