Allow 'help' options, in addition to 'help' command.
This is just syntax sugar.

Also, given the list of option tuples from gnu_getopt,
we convert the pairs to a dict for easier use.

Also also, 'optupi' is a masterful pun on 'opt tuples' and 'octopi.'
If you ever dare to change that variable name, my ghost will haunt you
into eternity.