Add min and max levels for each game.
This is based purely on what's in the database at the moment.
I assume there are no unreached levels in our games.
Note, 'level' is a fuzzy concept for Lakeland.
John has something rigged up in the code to calculate the sliding windows,
I just put in -1 for the levels since the actual db values have no
meaning.