Major rework of the GameTable code.
This was started in earlier commits,
finished here.
When creating a GameTable, you can either pass in all the data yourself,
or you can use factory-style functions to build the table.
The functions available are FromDB and FromCSV.
The DB function requires a database connection and settings, plus the
request object.
The CSV method requires a pandas dataframe.
Truth be told, there's a lot of code in OpenGameData that would be
unnecessary if I'd known pandas from the start.