Move some code back to action, and add a second job.
Moved code to get date to action, and moved vpn connection back too.
Removed the upload of artifacts, long-term we won't use that,
and now we're down to just trying to get rsync working.
Lastly, used a little copy-paste so we have two jobs, to test if it'll
run them in parallel nicely.