Remove MAX_SESSIONS and MIN_MOVES.
MIN_MOVES was never implemented, and MAX_SESSIONS was only in for
database exports.
To my knowledge, we never actually used the MAX_SESSIONS limit, just
left it at the default of exporting all sessions in a date range.
So, removing these for now, and we can decide later if they were
actually desirable.
I'm thinking the answer is probably no, and this sort of filtering
should be done during actual data analysis.