Minor fixes to last work on Request, and remove a lot of dead code.
Most of the removed code was from old architecture where we had
different requests for different combinations of range type and source,
but not all possible combinations.

With refactoring, we now give a Request an Interface, an ExporterRange,
and ExporterFiles.
The Interface and ExporterRange will let us get session IDs from any
source and based on date or ID ranges, in theory.