Reorganize CalculateAggregateFeatures functions.
So, new miniextractors each have a function to calculate out current
value, which will handle the old 'calculate aggregates' functionality.
However, this is kinda needed for compatibility with old extractors.
So we have a useless version in Extractor, and an override in
LegacyExtractor.
The override calls an abstract function, which the LegacyExtractor
subclasses each need to override.