Make a couple functions static/class-level functions.
In particular, we want to have feature and event dependencies as class
methods, since it shouldn't depend on a particular instance, but we
still need to be able to call private function through child class,
which requires us to have class var.
We'd also like Max/MinVersion and AvailableModes as statics.