Give GameSchema a 'legacy mode' that ensures we don't reload extractor many times.
Basically, by default the Registry would loop over every feature and
instantiate an instance of the LegacyExtractor every time.
Now, we ensure that if there is a 'legacy' item in the feature
dictionary, the GameSchema will flip to 'legacy mode' and will only
report a single 'legacy' feature as enabled.
This way, we don't get a gazillion instances of the monolithic
extractor.

Also, set up the LegacyLoader to return 'None' when asked for an
Extractor class.