Rework some of the config items in GH actions to reflect latest changes.
Specifically, the config action doesn't try to set a db name, since
that's no longer in the config for a MySQL type source.
This commit also adds an action + workflow for 'legacy' sources,
basically it gives a way to set the source for a game to something other
than the default.