Replace old behavior for missing end date; now just do the single day.
Before, we treated a single date as the starting date,
and exported every day from then to the present.
This made a lot of sense 4 years ago when we had like a month or two of
data. Does not make any sense now, with years upon years to run if you
set an early enough starting date.