Move certain components out of 'schemas' into 'models'
So far, we've lazily referred to things like Event as a "schema,"
which, in a very general manner of speaking, they are.
But as things have developed, we've used the word "schema"
to refer to things we load from (and save to) JSON-like formats.
Events, on the other hand, are internal models used to transfer and operate on data.
Thus, they are better described as "models."

Similarly, we've got a bunch of enums in the "schemas" folder.
With all this in mind, we've moved these kinds of "schemas"
to their own "models" folder, separate from all the things that are subclassing Schema.

Everything is in one giant commit because, well, everything was gonna be
touched anyway, so one commit that modifies damn-near every file is
better than 4 such commits.