Turn last week, last month, etc. workflows back into actions
Just discovered that we can actually use custom actions to run other
actions, not sure how/why I always thought that was impossible.
Maybe it used to be, or maybe I was just always an idiot and/or used a
wrong path if/when I tried it before.

Anyway, clear out the testing action & workflow, and convert the export
workflows back to actions.