QuitOnCityFailure doesn't need a failure reason subfeature.
If we had to hardcode it, including it doesn't really help us.
Not sure that there is even a true difference in reasons for a
"city fail", if so we'll need to go back and actually find what
possibilities are and how to determine which occurred.

Lastly, get county name from EventData instead of GameState.