Further work to get more reasonable values from StopwatchTimer
Turns out we need to account for more corner cases.
Rather than a single 'reset' event, we need a list of events that count
as 'resets', as well as a list of events to be ignored in general.
The 'resets' act like an 'end' event, but also are meant to be ignored
in time counting.
This way, we have start and end for normal gameplay, ignore for anything
we don't care to include in the counts, i.e. events that only occur when
players re-open the game after time away, but before they're back to
playing the game, and reset events in case the 'ignore' events coincide
with whatever mode/interaction we were recording being reset.
Also, added documentation of these params to the module comment block.
For future reference, need a good way to document these.