Mystery solved, earlier change was for wrong feature.
Earlier commit changed IdleSeconds -> Idle-Seconds in the Aqualab
template, but for the SessionDurationPerJob feature, which was
incorrect.
More recent commit, where the message indicated confusion because the
prior fix seemed to have been missed, was in fact the correct change,
for the previously-incorrect TimeInJournal config.

This commit, then, reverts the original, so that both features now have
their correct naming convention for the subfeatures matching between
their config and their modules' Subfeatures functions.

The two use opposite conventions, which is just confusing generally,
but between the modules and the configs, they're consistent now.